Squashed 2 commits.
[510da5fdaa06e78b5d46ad8ae94afa7b37fab21b] navigations

[4052e54c098978820542023dfdbb870205f5bc8e] env add mailtrap

+++ app/Http/Controllers/cmsbackend/TrashController.php
@@ -9,11 +9,12 @@ use App\Repositories\Contracts\LanguageRepositoryInterface;
 use App\Repositories\Contracts\FormRepositoryInterface;
 use App\Repositories\Contracts\PageRepositoryInterface;
 use App\Repositories\Contracts\ModuleRepositoryInterface;
+use App\Repositories\Contracts\NavigationRepositoryInterface;
 use Auth;
 
 class TrashController extends BackendController
 {
-    public function __construct(UserRepositoryInterface $user, RedirectRepositoryInterface $redirect, TranslationRepositoryInterface $translation, LanguageRepositoryInterface $language, FormRepositoryInterface $form, PageRepositoryInterface $page, ModuleRepositoryInterface $module)
+    public function __construct(UserRepositoryInterface $user, RedirectRepositoryInterface $redirect, TranslationRepositoryInterface $translation, LanguageRepositoryInterface $language, FormRepositoryInterface $form, PageRepositoryInterface $page, ModuleRepositoryInterface $module, NavigationRepositoryInterface $navigation)
     {
         parent::__construct();
         $this->user = $user;
@@ -23,6 +24,7 @@ class TrashController extends BackendController
         $this->form = $form;
         $this->page = $page;
         $this->module = $module;
+        $this->navigation = $navigation;
     }
 
     /**
@@ -39,6 +41,7 @@ class TrashController extends BackendController
         $forms = $this->form->paginatedFormsTrash();
         $pages = $this->page->paginatedPagesTrash();
         $modules = $this->module->paginatedModulesTrash();
+        $navigations = $this->navigation->paginatedNavigationsTrash();
         $this->breadcrumbs->addCrumb(__('Usunięte elementy'), '/cmsbackend/trash');
         return view('cmsbackend.trash.index')->with([
             'users' => $users,
@@ -48,6 +51,7 @@ class TrashController extends BackendController
             'forms' => $forms,
             'pages' => $pages,
             'modules' => $modules,
+            'navigations' => $navigations,
             'breadcrumbs' => $this->breadcrumbs,
             'pageTitle' => __('Usunięte elementy'),
             'is_active_nav' => 'trash'
@@ -79,11 +83,18 @@ class TrashController extends BackendController
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($module, $id)
+    public function destroy($module, $id, $type = '')
     {
         $statusmsg = __('Element usunięty');
 
-        $this->$module->destroy($id);
+        if($type == 'destroy') {
+            $this->$module->destroy($id);
+        } else {
+            $this->$module->update([
+                'status' => 4,
+                'who_updated' => Auth::id()
+            ], $id);
+        }
 
         return redirect()->route('trash')->with([
             'status-'.$module => $statusmsg,
